Add "Gaming" to Retro-Go menu item
As many kids (and adults) don't know what Retro-Go is,
I thought it good to add "Gaming" to the menu entry
so that it's more clear what it is.

First I considered "Retro-Go Games" but the games themselves
are not actually built by Retro-Go. It's the "Retro Gaming"
*experience* that is provided by Retro-Go. That's why I went with
"Retro-Go Gaming" in the end.